- Fully working RV64
- ISA switching
- Multicore support (SMP)
- RVVM Lib API
- Dynamic Device Tree generation
- Optimizations (~20% faster RAM, framebuffer caching)
- Poweroff/reset
- Precise interrupts
- Misaligned MMIO fixup
- More VM arguments & better arg parser (see rvvm -h)
- CMake build system
- Removed lots of legacy code
- Fixed Ethernet out-of-DMA corruption
- Fixed PS2 keyboard key repeat
- Fixed TLB pollution in IRQs/traps
- Fixed some race conditions
- Fixed >=4GB hard drives on Windows

+++ src/ringbuf.c
@@ -16,7 +16,7 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include <ringbuf.h>
+#include "ringbuf.h"
 #include <utils.h>
 #include <string.h>
 
@@ -46,7 +46,7 @@ bool ringbuf_is_empty(struct ringbuf *rb)
 bool ringbuf_put(struct ringbuf *rb, void *data, size_t len)
 {
 	if (ringbuf_get_free_spc(rb) < len) {
-		rvvm_warn("Overflow in ringbuf %p! size=%lu, consumed=%lu, len=%lu", rb, rb->size, rb->consumed, len);
+		rvvm_warn("Overflow in ringbuf %p! size=%u, consumed=%u, len=%u", rb, (uint32_t)rb->size, (uint32_t)rb->consumed, (uint32_t)len);
 		return false;
 	}
